Extract Graphviz logic into separate class
